document plugins and extract built-ins into them
This will clean up the built-in menu quite a bit!
