+++ docs/using-esc/plugins.rst
@@ -1,0 +1,54 @@
+=======
+Plugins
+=======
+
+esc bundles only a small number of built-in operations
+so you can decide what operations will be useful to you
+and avoid staring all day at operations you never use.
+Any remaining operations you need can be added via plugins.
+
+
+Plugin location
+===============
+
+Plugins are Python modules (``.py`` files)
+stored in the ``plugins`` subdirectory of your esc configuration directory.
+By default your esc configuration directory is ``~/.esc``,
+where ``~`` represents your home or user directory.
+If this directory doesn't exist but ``~/.config/esc`` does,
+that is used instead.
+
+
+Installing plugins
+==================
+
+To install a plugin,
+you simply copy its ``.py`` file into your plugins directory
+(see :ref:`Plugin location`).
+You can create the ``.esc`` and/or ``plugins`` directory if they don't exist.
+The next time you start esc, the plugins will be loaded.
+
+.. warning::
+    esc plugins can execute arbitrary Python code on your computer,
+    which could include malicious code,
+    so you should not install esc plugins that you do not trust.
+
+.. tip::
+    Plugins are loaded
+    and their operations placed in the :guilabel:`Commands` window
+    in alphabetical order by their filename.
+    To control the order, you can prefix their filenames with numbers,
+    e.g., ``01_trig.py``, ``02_log.py``.
+
+
+Finding plugins
+===============
+
+You can write your own plugins
+(see the :ref:`Developer Manual`)
+or get plugins from someone else.
+Some plugins providing common features like trig and log functions
+are available for download in the esc repository
+under ``esc-plugins`` (see `here`_).
+
+.. _here: https://github.com/sobjornstad/esc/tree/master/esc-plugins
